requirements.txt
pytest
allure-pytest
allure-python-commons